feat: introduce v1beta1/common package
